+++ CleanSwift/Source/Assembler.xctemplate/___FILEBASENAME___SceneAssembler.swift
@@ -2,10 +2,8 @@ import UIKit
 
 /// Responsible for assembling the scene and dependency injection.
 class ___VARIABLE_sceneName___SceneAssembler: ___VARIABLE_sceneName___SceneAssembling {
-    private(set) var isSceneAssembled = false
-
     /// Assembles the scene by creating and assigning all components of the scene and its dependencies if not done already.
-    func assembleIfNeeded(_ viewController: ___VARIABLE_sceneName___SceneViewController) {
+    func assembleIfNeeded(_ viewController: ___VARIABLE_sceneName___SceneViewController, isAssembled: inout Bool) {
         guard !isSceneAssembled else { return }
         isSceneAssembled = true
