+++ CleanSwift/Source/Module.xctemplate/___FILEBASENAME___Module.swift
@@ -23,24 +23,6 @@ public enum ___VARIABLE_moduleName___Module {
     public static func defaultController() -> ___VARIABLE_moduleName___Controller { ___VARIABLE_moduleName___DefaultController() }
 }
 
-// MARK: - Module Controller (module access interface)
-
-public protocol ___VARIABLE_moduleName___RootSceneRouter {
-    func startScene() -> UIViewController?
-}
-
-/// An interface of an instance of the module controller allowing data access and loading/managing the scenes of the module
-public protocol ___VARIABLE_moduleName___Controller {
-    typealias SomeData = Data
-    //var dataUpdateHandler: ((SomeData) -> Void)? { get set }
-
-    /// Requests data from the local storage.
-    //func requestLocalData(completion: @escaping (SomeData?) -> Void)
-    /// Requests data by sending the request to the `___VARIABLE_moduleName___DataSource` if data soruce was set with the `___VARIABLE_moduleName___Module.configure(dataSource:)` method.
-    //func requestDataUpdate(completion: ((SomeData?) -> Void)?)
-    //func update(data: SomeData)
-}
-
 // MARK: - Module Data Source (data pulling)
 
 public protocol ___VARIABLE_moduleName___DataSource {
